requirements.txt
Django==3.2
gunicorn==20.1.0
pillow==9.2.0
whitenoise==6.2.0
python-dotenv==0.21.0
social-auth-app-django==5.0.0
django-bootstrap-v5==1.0.11
django-crispy-forms==1.14.0
importlib-metadata==4.12.0
dj-database-url==1.0.0
psycopg2==2.9.3
boto3==1.24.88
django-storages==1.13.1